Upgrade to webpack 5 and remove watch option polling issue on windows abd remove obsolete webpack config rules
